Fix cache by registering it on instrumentation
NextJS variables are short lived per API request. Therefore, if multiple users try to access the app, the API Client will be called for each single one of them and the caching would not work.

To fix this issue, the cache is now registered using the instrumentation feature. This way, the cache will be available for the entire server lifetime